24 : Average Calories
I have to say I am pretty pleased with this one.  The main challenge was fairly easy, but the additional challenge offered the chance to practice ARGS.  This way you can pass in any number of lists and the function will work regardless and gets no more complex, even for a million lists!