25 : All The Same
These challenges are defiantly stepping up in the skills needed to complete them efficiently.  Its OK though Google and Stack Overflow are there to help!

For the core task the first challenge I found was that strings, lists ans tuples needed different algorithms for the comparison. I discovered the instance method to separate them out and I have run the test on all types in the main list.

The additional challenge shows the real power of Python for this kind of manipulation with split, reverse and join methods coming together!