32 :: Password Validator
A step up now, with the introduction of Regular Expressions, regex.  Using them is pretty easy, they return what has been found or none.  Constructing them is difficult to learn, but simple enough when you have done it a few times!  Personally, anything complex and I GPT or Bard it!