33 : List Intersection
Although the primary task asks for list comprehensions it seemed much more pythonic to me to do this in one, much more readable line.  So I did.

The extension task was a bit hard to understand, but ultimately introduces the time module.  There are other ways according to StackOverflow, but I couldn't get timeit to work.

The answer is .......lists.