37 : Count The Vowels
I think I have this one right, it is not clear in the challenge that by counting upper and lower case vowels a and A counts as 2 hits, I have assumed it is.  It would be easy enough to change with by using a capitalise method if you had to though.