39 : Password Generator
This was quite an interesting challenge, I first build the generator with a FOR loop, picking a random item from a set, but it felt too clunky.  So with some research I found the string module and eventually put it in to list comprehension to make the function really short and i think still really readable.