41 :: Words With Vowels
The entry task is pretty easy.  I added in a method to remove apostrophes (could be extended to other special characters) as after I did the test phrase I copied and pasted larger text strings and ' words broke it.  Line breaks in the text mean it will only look at the first complete line.

The extension activity is our first look at classes, it does open our eyes to the power if OOP for code reuse.