49 :: Create A Database
This is an intro the the world of databases, though it is fairly basic stuff it does help you understand how using the SQLite3 module is quite different to using most other modules.  The concept to embrace is that you create an object that connects to the DB file (conn) and then are instructing your python script to type SQL commands via the cursor command.  

Once I go tm mind round this it became easier to set the commands as variables.  If I had been the author I would have set an extension task to turn this into CRUD database!